==> Code/03-Integration_Clustering.R <==
library(Seurat)
library(SingleCellExperiment)
library(ggplot2)
library(patchwork)
library(dplyr)
library(clustree)
library(DoubletFinder)
library(DropletUtils)
library(biomaRt)
library(shadowtext)

# Read in sample objects generated by DoubletFinder processing
healthy <- readRDS("path/to/healthy_SoupX-DF.RDS")
eau1 <- readRDS("path/to/EAU1_SoupX-DF.RDS")
eau2 <- readRDS("path/to/EAU2_SoupX-DF.RDS")




# Make list of samples to pass through functions
samples <- c(healthy, eau1, eau2)

# Perform SCTransform and integration ----
features <- SelectIntegrationFeatures(object.list = samples, nfeatures = 3000)
samples <- PrepSCTIntegration(object.list = samples, anchor.features = features)

anchors <- FindIntegrationAnchors(object.list=samples, normalization.method="SCT", anchor.features = features)
data <- IntegrateData(anchorset=anchors, normalization.method="SCT")

# Dimensionality reduction and clustering
data <- RunPCA(data, npcs = 75)
elb <- ElbowPlot(data, ndims=75)
elb

data <- RunUMAP(data, dims = 1:30) %>% FindNeighbors(dims = 1:30)

# Use several resolutions for input into clustree
data <- FindClusters(data, resolution=seq(0.3, 1.5, 0.1))

# Visualize with clustree to determine reasonable cluster resolution
clustree <- clustree(data, prefix="integrated_snn_res.",
                     node_colour="sc3_stability", 
                     node_size_range=c(4,10), 
                     edge_arrow=FALSE, 
                     show_axis=TRUE)

clustree

# Set cluster resolution
Idents(data) <- "integrated_snn_res.0.6"

# Visualise clustering
p1 <- DimPlot(data, label=T) + NoLegend()
p1
# Visualise integration
DimPlot(data, label=T, group.by = "orig.ident")


# Cluster annotation
ids <- c("Rods",
         "Rods",
         "Rods",
         "Rods",
         "Rod Bipolar Cells",
         "Cones",
         "Müller Glia",
         "Immune Cells",
         "Cones",
         "Cone Bipolar Cells",
         "Immune Cells",
         "Müller Glia",
         "Cones",
         "Cone Bipolar Cells",
         "Amacrine Cells",
         "Cone Bipolar Cells",
         "Cones",
         "Rod Bipolar Cells",
         "Cone Bipolar Cells",
         "Amacrine Cells",
         "Amacrine Cells",
         "Cone Bipolar Cells",
         "Cone Bipolar Cells",
         "Cone Bipolar Cells",
         "Cone Bipolar Cells",
         "Junk",
         "RPE")

names(ids) <- levels(data)
data <- RenameIdents(data, ids)

data[["cell_ids"]] <- Idents(data)

# Remove ambiguous cluster
data <- subset(data, idents="Junk", invert=TRUE)


# Re-split samples to pass through functions
samples <- SplitObject(data, split.by = "orig.ident")

# Perform SCTransform and integration ----
samples <- lapply(samples, function(x) SCTransform(x, method = "glmGamPoi", vars.to.regress = c("percent.mt", "percent.ribo")))

features <- SelectIntegrationFeatures(object.list = samples, nfeatures = 3000)
samples <- PrepSCTIntegration(object.list = samples, anchor.features = features)

anchors <- FindIntegrationAnchors(object.list=samples, normalization.method="SCT", anchor.features = features)
data <- IntegrateData(anchorset=anchors, normalization.method="SCT")

# Dim reduction and clustering
data <- RunPCA(data, npcs = 75)
elb <- ElbowPlot(data, ndims=75)
elb

data <- RunUMAP(data, dims = 1:40) %>% FindNeighbors(dims = 1:40)

# Use several resolutions for input into clustree
data <- FindClusters(data, resolution=seq(0.3, 1.5, 0.1))

# Visualize with clustree to determine reasonable cluster resolution
clustree <- clustree(data, prefix="integrated_snn_res.",
                     node_colour="sc3_stability", 
                     node_size_range=c(4,10), 
                     edge_arrow=FALSE, 
                     show_axis=TRUE)

clustree

# Set cluster resolution
Idents(data) <- "integrated_snn_res.0.7"

# Visualise clustering
p1 <- DimPlot(data, label=T) + NoLegend()
p1
# Visualise integration
DimPlot(data, group.by = "orig.ident", split.by = "orig.ident")

DefaultAssay(data) <- "RNA"
data <- NormalizeData(data) %>% ScaleData()

# Rename idents
ids <- c("Rods",
         "Rods",
         "Rods",
         "Rods",
         "Rod Bipolar Cells",
         "Cones",
         "Müller Glia",
         "Immune Cells",
         "Cones",
         "Cone Bipolar Cells",
         "Cones",
         "Müller Glia",
         "Immune Cells",
         "Cones",
         "Amacrine Cells",
         "Cone Bipolar Cells",
         "Rod Bipolar Cells",
         "Cone Bipolar Cells",
         "Amacrine Cells",
         "Amacrine Cells",
         "Cone Bipolar Cells",
         "Cone Bipolar Cells",
         "Cone Bipolar Cells",
         "Cone Bipolar Cells",
         "Cone Bipolar Cells",
         "Cone Bipolar Cells",
         "Cone Bipolar Cells",
         "Rods",
         "RPE",
         "Immune Cells")

names(ids) <- levels(data)
data <- RenameIdents(data, ids)

# Update cell_ids metadata column
data[["cell_ids"]] <- Idents(data)          


# Immune Cells ----
ics <- subset(data, idents="Immune Cells")

# Run standard SCTransform pipeline on ICs
DefaultAssay(ics) <- "RNA"
ics <- SCTransform(ics, method="glmGamPoi", vars.to.regress = c("percent.mt", "percent.ribo"))

ics <- RunPCA(ics, npcs = 50)
ElbowPlot(ics, ndims = 50)

DefaultAssay(ics) <- "SCT"
ics <- RunUMAP(ics, dims = 1:39)
ics <- FindNeighbors(ics, dims = 1:39)
ics <- FindClusters(ics, resolution = 1.3)

clustree(ics, prefix="SCT_snn_res.",
         node_colour="sc3_stability", 
         node_size_range=c(4,10), 
         edge_arrow=FALSE, 
         show_axis=TRUE)


Idents(ics) <- "SCT_snn_res.1.3"

DimPlot(ics, label=T) + NoLegend()

DefaultAssay(ics) <- "RNA"
ics <- NormalizeData(ics) %>% ScaleData()


ic_markers <- FindAllMarkers(ics, min.pct = 0.25, only.pos = T) %>% 
  group_by(cluster) %>% slice_max(n=10, order_by = avg_log2FC)


View(ic_markers)

ic_ids <- c("Microglia",
            "Naive CD4+ T cells",
            "Rogdi+ MPs",
            "CD8+ T cells",
            "Monocytes",
            "Tregs",
            "Th1 Cells",
            "NK cells",
            "Neutrophils",
            "pDCs",
            "Th17/gd T cells")

names(ic_ids) <- levels(ics)
ics <- RenameIdents(ics, ic_ids)

DimPlot(ics, label=T) + NoLegend()

# cell numbers               
table(Idents(ics))

#Removing Rogdi MPs and reprocessing
ics <- subset(ics, idents = "Rogdi+ MPs", invert=T)

DefaultAssay(ics) <- "RNA"
ics <- SCTransform(ics, method="glmGamPoi", vars.to.regress = c("percent.mt", "percent.ribo"))

ics <- RunPCA(ics, npcs = 50)
ElbowPlot(ics, ndims = 50)

DefaultAssay(ics) <- "SCT"
ics <- RunUMAP(ics, dims = 1:20)
ics <- FindNeighbors(ics, dims = 1:20)
ics <- FindClusters(ics, resolution = seq(0.3, 1.5,0.1))

clustree(ics, prefix="SCT_snn_res.",
        node_colour="sc3_stability",
        node_size_range=c(4,10),
        edge_arrow=FALSE,
        show_axis=TRUE)

Idents(ics) <- "SCT_snn_res.1.2"

DimPlot(ics, label=T) + NoLegend()

DefaultAssay(ics) <- "RNA"
ics <- NormalizeData(ics) %>% ScaleData()


ic_markers <- FindAllMarkers(ics, min.pct = 0.25, only.pos = T) %>% 
  group_by(cluster) %>% slice_max(n=10, order_by = avg_log2FC)

View(ic_markers)

ic_ids <- c("Microglia",
            "Naive CD4+ T cells",
            "CD8+ T cells",
            "Monocytes",
            "Tregs",
            "Th1 Cells",
            "NK cells",
            "Neutrophils",
            "Th17/gd T cells",
            "pDCs")

names(ic_ids) <- levels(ics)
ics <- RenameIdents(ics, ic_ids)

# Adding immune cell subcluster as subcolumn to full integrated data object                
data$ic_cluster <- as.character(Idents(data))
data$ic_cluster[Cells(ics)] <- paste(Idents(ics))

# Save final integrated data object for easier loading in future 
saveRDS(data, file="integrated_data.RDS")
