==> Code/06B-cellphonedb-analysis.ipynb <==
from cellphonedb.src.core.methods import cpdb_statistical_analysis_method
from cellphonedb.utils import db_utils
import os
import pandas as pd
import sys
import ktplotspy as kpy
import anndata as ad
import matplotlib.pyplot as plt
import seaborn as sn

db_utils.download_database("cpdb/db", "v4.1.0")

cpdb_file_path="cpdb/db/cellphonedb.zip"
meta_file_path="cpdb/CPDB/eau1_metadata.txt"
counts_file_path="cpdb/CPDB/eau1.h5ad"
output_path="results"

meta = pd.read_csv(meta_file_path, sep="\t")
adata = ad.read_h5ad(counts_file_path)

deconvoluted, means, pvalues, significant_means = cpdb_statistical_analysis_method.call(
    cpdb_file_path=cpdb_file_path,
    meta_file_path=meta_file_path,
    counts_file_path=counts_file_path,
    counts_data="hgnc_symbol",
    output_path=output_path,
    separator="|",
    threshold=0.1,
    result_precision=3,
    output_suffix="EAU01"
)

m = pd.read_csv("results/statistical_analysis_means_EAU01.txt", sep="\t")
p = pd.read_csv("results/statistical_analysis_pvalues_EAU01.txt", sep="\t")
d = pd.read_csv("results/statistical_analysis_deconvoluted_EAU01.txt", sep="\t")

ints = kpy.plot_cpdb_heatmap(adata=adata,
                     pvals=p,
                     celltype_key="ic_cluster",
                     cmap="viridis", 
                     return_tables = True)

network = ints["count_network"]
total_ints = ints["interaction_count"]
edges = ints["interaction_edges"]

network.to_csv("results/CPDB_network_EAU01.txt", sep="\t")
total_ints.to_csv("results/CPDB_total-interactions_EAU01.txt", sep="\t")
edges.to_csv("results/CPDB_edges_EAU01.txt", sep="\t")

cpdb_file_path="cpdb/db/cellphonedb.zip"
meta_file_path="cpdb/CPDB/eau2_metadata.txt"
counts_file_path="cpdb/CPDB/eau2.h5ad"
output_path="results"

meta = pd.read_csv(meta_file_path, sep="\t")

adata = ad.read_h5ad(counts_file_path)

deconvoluted, means, pvalues, significant_means = cpdb_statistical_analysis_method.call(
    cpdb_file_path=cpdb_file_path,
    meta_file_path=meta_file_path,
    counts_file_path=counts_file_path,
    counts_data="hgnc_symbol",
    output_path=output_path,
    separator="|",
    threshold=0.1,
    result_precision=3,
    output_suffix="EAU02"
)
